% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classifyDirectSpecies.R
\name{classifyDirectSpecies}
\alias{classifyDirectSpecies}
\title{Classify direct species PCR}
\usage{
classifyDirectSpecies(
  sub_directory,
  project,
  assay = "direct_species",
  plate_folder
)
}
\arguments{
\item{sub_directory}{name of folder within RProj where you are working}

\item{project}{name of the project folder you specified in startNewProject()}

\item{assay}{if this is direct PCR you should specify \code{"direct_species"}}

\item{plate_folder}{the name of the folder where your direct PCR data is stored}
}
\value{
list object with plots and raw data
}
\description{
Classify direct species PCR
}
\examples{
\dontrun{classifyDirectSpecies(sub_directory = "lab-molecular", project = "project_name", assay = "nested_species", plate_folder = "plate1")}
}
