% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classifyNestedSpecies.R
\name{classifyNestedSpecies}
\alias{classifyNestedSpecies}
\title{Classify Nested qPCR data}
\usage{
classifyNestedSpecies(
  sub_directory,
  project,
  assay = "nested_species",
  plate_folder
)
}
\arguments{
\item{sub_directory}{name of folder within RProj where you are working}

\item{project}{name of the project folder you specified in startNewProject()}

\item{assay}{if this is nested PCR you should specify \code{"nested_species"}}

\item{plate_folder}{the name of the folder where your nested PCR data is stored}
}
\value{
list object with plots and raw data
}
\description{
Classify Nested qPCR data
}
\examples{
\dontrun{classifyNested(sub_directory = "lab-molecular", project = "project_name", assay = "nested_species", plate_folder = "plate1")}

}
