% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classifyScreening.R
\name{classifyScreening}
\alias{classifyScreening}
\title{Classify screening results}
\usage{
classifyScreening(project, assay = "screening", plate_folder)
}
\arguments{
\item{project}{the name given during project setup (that is the same as within results or data)}

\item{assay}{is this screening, nested_species or direct_species}

\item{plate_folder}{what is the name of the folder where the results are saved}
}
\value{
a list of results to be accessed to find plots, data etc
}
\description{
Classify screening results
}
