% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createPCRdatabase.R
\name{createPCRdatabase}
\alias{createPCRdatabase}
\title{Create one database of all the results generated so far and stored in the respective results folders}
\usage{
createPCRdatabase(sub_directory, project, species_type)
}
\arguments{
\item{sub_directory}{name of sub_directory folder relative to RProj}

\item{project}{project name}

\item{species_type}{have the results been generated with nested or direct species}
}
\value{
an excel spreadsheet stored in the merged_database folder
}
\description{
Create one database of all the results generated so far and stored in the respective results folders
}
\examples{
\dontrun{createPCRdatabase(sub_directory = "lab-molecular", project = "test_project", species_type = "nested_species")}
}
