% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directSpeciesReport.R
\name{directSpeciesReport}
\alias{directSpeciesReport}
\title{Create report for direct species PCR}
\usage{
directSpeciesReport(
  sub_directory,
  project,
  assay = "direct_species",
  plate_folder,
  file_type = ".pdf"
)
}
\arguments{
\item{sub_directory}{subdirectory you are working within}

\item{project}{name of the project}

\item{assay}{if direct PCR you should specify "direct_species"}

\item{plate_folder}{the name of the folder where you are storing your data}

\item{file_type}{specify either ".html" or ".pdf" according to what output format you would like}
}
\value{
a pdf and excel spreadsheet saved in results folder
}
\description{
Create report for direct species PCR
}
\examples{
\dontrun{directSpeciesReport(sub_directory = "lab-molecular",project = "project_name",  assay = "direct_species", plate_folder ="example-directspecies-plate1", file_type = ".pdf" )}
}
