% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importQuantStudio.R
\name{importQuantStudio}
\alias{importQuantStudio}
\title{import files for QuantStudio}
\usage{
importQuantStudio(INPUT_PLATELAYOUT, INPUT_QUANTSTUDIO)
}
\arguments{
\item{INPUT_PLATELAYOUT}{an excel file in plate layout format}

\item{INPUT_QUANTSTUDIO}{the excel file exported from the quantstudio}
}
\value{
a list
}
\description{
import files for QuantStudio
}
\examples{
\dontrun{importQuantStudio(INPUT_PLATELAYOUT = INPUT_PLATELAYOUT, INPUT_QUANTSTUDIO = INPUT_QUANTSTUDIO)}
}
