% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nestedSpeciesReport.R
\name{nestedSpeciesReport}
\alias{nestedSpeciesReport}
\title{Create report for nested species PCR}
\usage{
nestedSpeciesReport(
  sub_directory,
  project,
  assay = "nested_species",
  plate_folder,
  file_type = ".pdf"
)
}
\arguments{
\item{sub_directory}{subdirectory you are working within}

\item{project}{name of the project}

\item{assay}{if nested PCR you should specify "nested_species"}

\item{plate_folder}{the name of the folder where you are storing your data}

\item{file_type}{specify either ".html" or ".pdf" according to what output format you would like}
}
\value{
a pdf and excel spreadsheet saved in results folder
}
\description{
Create report for nested species PCR
}
\examples{
\dontrun{nestedSpeciesReport(sub_directory = "lab-molecular",project = "project_name",  assay = "nested_species", plate_folder ="example-nestedspecies-plate1", file_type = ".pdf" )}
}
