% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/screeningReport.R
\name{screeningReport}
\alias{screeningReport}
\title{Generate a report of the screening experiment}
\usage{
screeningReport(
  sub_directory,
  project,
  assay = "screening",
  plate_folder,
  file_type = ".pdf"
)
}
\arguments{
\item{sub_directory}{specify the name of the sub_directory you save your data in, if working within a subfolder}

\item{project}{same as the project name you gave in setup}

\item{assay}{if generating screening results, this shoudl be "screening"}

\item{plate_folder}{the name of the folder where your two key data files are stored}

\item{file_type}{".pdf" or ".html" depending on the output file you want}
}
\value{
pdf or html file of results
}
\description{
Generate a report of the screening experiment
}
\examples{
\dontrun{screeningReport(sub_directory = "lab-molecular", project = "example",
assay = "screening", plate_folder = "plate1", file_type = ".pdf")}
}
