% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specifyFolder.R
\name{specifyFolder}
\alias{specifyFolder}
\title{Specify file location}
\usage{
specifyFolder(project, assay, plate_folder)
}
\arguments{
\item{project}{the name given during project setup (that is the same as within results or data)}

\item{assay}{is this screening, nested_species or direct_species}

\item{plate_folder}{what is the name of the folder where the results are saved}
}
\value{
a vector with relative paths from R proj, first the plate map, second the Quantstudio export file
}
\description{
Specify file location
}
\examples{
\dontrun{specifyFolder("example", "screening", "plate2-repeat")}
}
