% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/screeningReport.R
\name{screeningReport}
\alias{screeningReport}
\title{Generate a report of the screening experiment}
\usage{
screeningReport(project, assay, plate_folder, file_type = ".pdf")
}
\arguments{
\item{project}{same as the project name you gave in setup}

\item{assay}{if generating screening results, this shoudl be "screening"}

\item{file_type}{".pdf" or ".html" depending on the output file you want}

\item{plaste_folder}{the name of the folder where your files are stored}
}
\value{
pdf or html file of results
}
\description{
Generate a report of the screening experiment
}
\examples{
\dontrun{screeningReport("example", "screening", "plate1", ".pdf")}
}
