% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/startNewProject.R
\name{startNewProject}
\alias{startNewProject}
\title{Generate folder structure to save results}
\usage{
startNewProject(sub_directory = "", project, species_type)
}
\arguments{
\item{sub_directory}{if working within a folder at the root, specify this here. if not, leave blank}

\item{project}{project name to use in folder structure}

\item{species_type}{either 'nested_species' or 'direct_species' depending on the type of species protocol to be use for the project}
}
\value{
folder structure at current working directory
}
\description{
Generate folder structure to save results
}
\examples{
\dontrun{startNewProject(sub_directory = "lab-molecular",project = "example",  species_type = "nested_species")}

}
