% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/startNewProject.R
\name{startNewProject}
\alias{startNewProject}
\title{Generate folder structure to save results}
\usage{
startNewProject(project_name, species_type)
}
\arguments{
\item{project_name}{project name to use in folder structure}

\item{species_type}{either 'nested' or 'direct' depending on the type of species protocol to be use for the project}
}
\value{
folder structure at current working directory
}
\description{
Generate folder structure to save results
}
\examples{
\dontrun{startNewProject("STATEM", "nested")}

}
